# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Customer Manager API",
        "version": "0.1.0"
    },
    "paths": {
        "/customers": {
            "get": {
                "tags": [
                    "customers"
                ],
                "summary": "List all customers",
                "description": "List all customers",
                "operationId": "listAllCustomer",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error: Unprocessable Content"
                    }
                }
            },
            "post": {
                "tags": [
                    "customers"
                ],
                "summary": "Add a new customer",
                "description": "",
                "operationId": "addCustomer",
                "requestBody": {
                    "description": "Customer object that needs to be added",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCustomer"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Error: Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error: Unprocessable Content"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CreateCustomer": {
                "required": [
                    "name",
                    "customers"
                ],
                "properties": {
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "document": {
                        "type": "string"
                    },
                    "birth_name": {
                        "type": "string"
                    },
                    "phone_number": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Customer": {
                "required": [
                    "name",
                    "customers"
                ],
                "properties": {
                    "uuid": {
                        "type": "string"
                    },
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "document": {
                        "type": "string"
                    },
                    "birth_name": {
                        "type": "string"
                    },
                    "phone_number": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}